requirements.txt
django==5.2.3
djangorestframework==3.16.1
python-dotenv==1.1.1
django-ckeditor==6.7.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
